Store pointer bytes in pointers, don't transmute
Round-trips of pointers through integers are extremely problematic for a
memory model to accomodate, and if done by a transmute are basically
impossible. The problem is that a compiler doing provenance-based
optimizations can't keep track of which memory has had its address
preserved, so with ptr-int-ptr transmutes your choices are basically no
provenance optimizations or UB. And we like optimizations. These
operations are not yet formally declared to be UB, but it's hard to
imagine an alternative.